+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-08-18 18:42:47
-// Commit: feat(i18n): add async i18n stream provider
+// Updated at: 2026-08-19 09:18:36
+// Commit: feat(i18n): integrate custom telemetry logger in i18n
 
 class i18nComponentState {
-  final String commitRef = '3713';
-  final String lastMessage = 'feat(i18n): add async i18n stream provider';
-  final DateTime timestamp = DateTime.parse('2026-08-18T18:42:47+05:00');
+  final String commitRef = '3731';
+  final String lastMessage = 'feat(i18n): integrate custom telemetry logger in i18n';
+  final DateTime timestamp = DateTime.parse('2026-08-19T09:18:36+05:00');
 }